refactor(mongo): extract shared @dbx-app/mongo-shell package
Centralize shell JSON helpers and thin aggregate parse in a pure workspace
package so desktop and node-core share one implementation without helper
copies or relative path imports into node-core.
