+++ packages/mongo-shell/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "@dbx-app/mongo-shell",
+  "version": "0.1.0",
+  "private": true,
+  "description": "Pure MongoDB shell JSON/argument parsing shared by desktop and node-core",
+  "license": "Apache-2.0",
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js",
+      "default": "./dist/index.js"
+    }
+  },
+  "files": [
+    "dist",
+    "src"
+  ],
+  "scripts": {
+    "build": "tsc",
+    "prepare": "tsc"
+  },
+  "devDependencies": {
+    "typescript": "^5.8.3"
+  },
+  "engines": {
+    "node": ">=22.13.0"
+  }
+}
