+++ packages/mongo-shell/src/index.ts
@@ -1,0 +1,20 @@
+export {
+  chainedMethodCallPattern,
+  escapeRegExp,
+  findChainedMethodCallIndex,
+  findMatchingParen,
+  hasUnclosedMongoDelimiters,
+  normalizeJsonArgument,
+  parseCollectionMethodTarget,
+  parseMongoObjectArgument,
+  quoteUnquotedObjectKeys,
+  splitTopLevel,
+  trimMongoOuterComments,
+} from "./json.js";
+
+export {
+  describeMongoCommandParseFailure,
+  MONGO_SHELL_COMMAND_HINT,
+  parseMongoAggregateCommand,
+  type MongoAggregateCommand,
+} from "./aggregate.js";
